pyproject.toml
[project]
name = "robo-advisor"
version = "0.1.0"
description = "100% local multi-asset Robo-Advisor for Vietnam market with AI/RL trading"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    # Core
    "pandas>=2.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "scipy>=1.11",
    "pydantic>=2.0",
    # Market Data
    "vnstock>=0.3",
    # Streaming
    "kafka-python>=2.0",
    # Storage
    "clickhouse-driver>=0.2.6",
    # UI
    "streamlit>=1.28",
    "plotly>=5.0",
    # Notifications
    "requests>=2.31",
    "fpdf2>=2.7",
    "schedule>=1.2",
    # Async
    "aiohttp>=3.9",
]

[project.optional-dependencies]
ml = [
    "stable-baselines3>=2.0",
    "gymnasium>=0.29",
    "transformers>=4.35",
    "torch>=2.1",
    "mlflow>=2.10",
    "evidently>=0.4",
    "scikit-learn>=1.3",
]
broker = [
    "requests-oauthlib>=1.3",
]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
